deepfake_detection_platform/requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn>=0.22.0
streamlit>=1.22.0
python-multipart>=0.0.6
pydantic>=1.10.7
aiofiles>=23.1.0

# Database
aiosqlite>=0.18.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.7.0.72
scipy>=1.10.1
numpy>=1.24.3

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1

# Model Support
# Uncomment based on models you want to use
# tensorflow>=2.12.0
# torch>=2.0.0
# onnxruntime>=1.14.1

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1
altair>=5.0.0
pandas>=2.0.1

# Utilities
python-dotenv>=1.0.0
python-magic>=0.4.27
reportlab>=3.6.13
tqdm>=4.65.0

# Development Tools
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
